Dojo: Restructure rooms (#63)
Swap rooms (r11 -> l1, r3 -> l3, r9 -> l4)

closes #60
