+++ dojo-dungeon/src/dojo/rooms/builder/RoomBuilder.java
@@ -7,17 +7,17 @@ import core.utils.components.path.IPath;
 import dojo.rooms.LevelRoom;
 import dojo.rooms.Room;
 import dojo.rooms.level_1.Room1;
+import dojo.rooms.level_1.Room11;
 import dojo.rooms.level_1.Room2;
-import dojo.rooms.level_1.Room3;
 import dojo.rooms.level_2.Room4;
 import dojo.rooms.level_2.Room5;
 import dojo.rooms.level_2.Room6;
+import dojo.rooms.level_3.Room3;
 import dojo.rooms.level_3.Room7;
 import dojo.rooms.level_3.Room8;
-import dojo.rooms.level_3.Room9;
 import dojo.rooms.level_4.Room10;
-import dojo.rooms.level_4.Room11;
 import dojo.rooms.level_4.Room12;
+import dojo.rooms.level_4.Room9;
 import java.util.ArrayList;
 import java.util.HashMap;
